refactor(user_routes.py): move login, signup, token, profile related routes to user routes
