+++ vidwiz/templates/login.html
@@ -49,7 +49,7 @@
           const errorDiv = document.getElementById('login-error');
           errorDiv.style.display = 'none';
           errorDiv.textContent = '';
-          const resp = await fetch('/login', {
+          const resp = await fetch('/user/login', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify({ username, password })
