+++ vidwiz/templates/signup.html
@@ -70,7 +70,7 @@
         submitButton.textContent = 'Signing Up...';
         
         try {
-          const response = await fetch('/signup', {
+          const response = await fetch('/user/signup', {
             method: 'POST',
             headers: {
               'Content-Type': 'application/json',
